Some fairly extensive changes to the test rig.
There are two conceptual changes, but they are two textually entangled
to be easily checked in separately :-(

The first change is allowing a test to skip, by raising a
pypy.tool.test.TestSkip exception.

test.objspace() raises TestSkip if it is asked for a specific flavour
of object space and that differs from the specific flavour of object
space that is being tested (as controlled by command-line arguments).

This change is all of the changes to interpreter/unittest_w.py and
some of the work in tool/test.py.

The other change is a new option (-k, for no good reason other than -i
already being taken) that drops you into a interactive session when a
test run has failures.  From this you can list the errors ('l'), print
tracebacks ('tb') and drop into a pdb session for each failure ('d').
Blank lines and 'q' quit.

I've also added a sketal PyPy-specialized subclass of pdb.Pdb.


git-svn-id: http://codespeak.net/svn/pypy/trunk/src@1103 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada