An attempt at an app-level import which can handle package imports.
It passes all regression tests with TrivialObjSpace, 
but it's commented out at the moment because it fails for StdObjSpace.

It also needs regression tests of its own, to make sure it's working
correctly.

There is also the issue of namespace pollution with imp_* items in 
builtins - they're there for an app/interpreter level bridge. 

(This is a rush checkin to have it availible for the sprinters - 
who should feel free to trash it if they come up with something
better.)

git-svn-id: http://codespeak.net/svn/pypy/trunk/src@2316 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada