Generate Pyrex class definitions.
This is for attributes only, there is no support for methods, and it is not
clear if there should be.  If we keep C in mind then we should probably
generate global functions only, and introduce explicit tricks to emulate
virtual methods -- though I believe that Pyrex can do precisely that, so maybe
we should just go ahead and emit methods.



git-svn-id: http://codespeak.net/svn/pypy/trunk/src@2986 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada