Gateway cleanups: trying to remove interpreter-level ** arguments.
Now interp2app(f) no longer accepts interp-level functions with **.
Instead, the signature of f should either end with a * argument, or with an 
argument called __args__ (hum).  In the latter case, the app-level version of 
f appears to have a * and a ** argument, and the interp-level version will get 
all that nicely collected in an pypy.interpreter.argument.Arguments instance.

Hum, the goal was to solve the problem in the failing test, but it still 
fails.  See mail to pypy-dev "fun with functions".



git-svn-id: http://codespeak.net/svn/pypy/trunk/src@5309 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada