Urgh, another huge check-in.  This one introduces classes for various kinds of representations, and put them to the new genc_repr.py.  This is a clean-up of the conversion mess.  Now the representation classes, subclassing CRepr, provide a convert_to() method called when a value of some type (i.e. with some repr) needs to be converted to a value of some other, given type.
To support this model, the operations have a "cost" and typer.py will
select, among the variants of an operation, the one with the less cost
(including the cost of conversions).  The operation() method is (a bit)  
simpler now.

genc_typeset.py is much more readable now: the three huge functions have
been split up in genc_repr.py and in a set of build_OP_XXX() methods.  
Automatic chaining of conversion is gone again.


git-svn-id: http://codespeak.net/svn/pypy/trunk/src@6372 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada