Support for C function pointers.  Calling other C functions is now a special case of calling a C function pointer.  If you really know C, you'll remember the strange syntax of its type declarations, e.g. how to declare a function that returns a pointer to another function...  This is easier with typedefs to give a name to a type "pointer to function with such-and-such signature", but after some hesitations this is now done without typedefs using the approach outlined in genc_repr.cdecl().
A masochistic C nerd might consider enabling the declaration of several
variables of the same type in a single line:

   PyObject *v1, *v2, *v3;

This is really messy because the "*" must be repeated (or more precisely, 
attached to the name instead of the type).  Currently we just pretend the type 
is "PyObject*" and the names "v1", "v2", "v3", etc.



git-svn-id: http://codespeak.net/svn/pypy/trunk/src@6385 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada