frozendict is dead, long live the freezable Cache class.
for any code reachable by the annotator/translator you need to/should 
use pypy.tool.cache.Cache().  When the annotator sees such a
cache it calls the "freeze()" method on it after which it cannot be
modified anymore and also becomes hashable as a side effect. 

A Cache has a getorbuild() method where you pass in a key, a builder
and a space. In case the key is not in the cache yet and the cache
is not frozen, builder(key, space) will be called to obtain the 
value. 

Various places now use Cache instead of plain dictionaries. 



git-svn-id: http://codespeak.net/svn/pypy/trunk/src@7360 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada