Code to flow identity information into code like
if x is None:
    # x is annotated as None in here
else:
    # but not here.

Tests for same.  Also, tests for the similar isinstance hacks which somehow
seem to have not been checked in yesterday (sorry about that).

To make the flow_identity_info test work, I had to hack the constructor
of SomeTuple somewhat.  If this is the wrong thing to have done, the test
can be written another way -- but this seems natural.


git-svn-id: http://codespeak.net/svn/pypy/trunk/src@7365 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada