Ignore blocked blocks during transformations.  This fixes a bunch of assertion errors if you are transforming a flow graph that contains blocked blocks.
This doesn't change anything in practice because if there are any blocked
blocks, the annotator raises an assertion error and no transformations are
performed.  Armin suggested that in some instances (e.g. calling a function
that always raises an exception) the blocked block should stay blocked and we
should proceed with the translation anyway.



git-svn-id: http://codespeak.net/svn/pypy/trunk/src@7527 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada