Right code, wrong place: bye bye CallableFactory, SomeCallable.
We're well on the way to 

svn mv factory.py bookkeeper.py

as the functionality of CallableFactory is now contained in the Bookkeeper
class.  SomeCallable is borged with SomePBC.

There's a lot of fairly mindless refactoring associated with this.

genpyrex is currently broken.  Probably shallow.

Add assertions that SomeBuiltins are never merged.

Teach annotator about list.pop, list.insert.  This possibly should be
a separate commit, but I'm too lazy for that.

diff gets rather the wrong idea about unaryop.py and factory.py.  What
can you do?


git-svn-id: http://codespeak.net/svn/pypy/trunk/src@7628 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada