copy the full Python 2.3.4 Lib hierarchy (including the regr-tests) into our svn/pypy/dist.
We probably want to move the "appspace"-hacks 
to this new directory so that we can easily
diff against the unmodified python2.3.4 
library and find out what we needed to change
in order to make PyPy work in a systematic 
manner. 

It is now probably a good idea to settle on going 
for python-2.3.4 implementations and features 
and only at some point later move to python-2.4. 



git-svn-id: http://codespeak.net/svn/pypy/dist@8639 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada