a subtle bug about exceptions. Holger tried to blame me about the generated exceptions module, but we could not find anything wrong. Finally we found out that everything was fine but the exit() function. This needs to use the com.ma version of raise. Why? CPython does it the same way, and by this way, normalize_exception does an extra de-tupelizing on the argument. :-) I think this is really an artifact, a wart in CPython. But well, right now we stay compatible.
By the way, it would be easy and really nice to allow
to raise any object,not just Exceptions. And in the case
of any object, we would *not* create tracebacks, but
treat that as an intended control flow exception.
Holger and I are dreaming of

raise 42

and

try:...
except int, val: # do something :-)

git-svn-id: http://codespeak.net/svn/pypy/dist@8681 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada