merge of the dist-interpapp branch where Samuele Armin and me worked for 9 days (summing up to a hundred commits, as you may have noticed in your inbox).
Major results (apart from little stuff here and 
there) are: 

- a new app2interp mechanism that is now indepedent from 
  the interp2app implementation 

- a new way to do "mixed modules" which contain applevel 
  and interplevel definitions.  We also ported the builtin
  and sys module and they now live as packages/directory 
  structures.  in pypy/module/sys2/__init__.py 
  and pypy/module/builtin/__init__.py you'll find dictionaries
  which specify where attributes are loaded from (on demand). 

  The old very hacky code for interplevel/applevel modules
  is gone. 

- in the course, interp2app was cleaned up a bit 
  and we added more type information/checks for GetSetProperties 
  and other interp-level functions. 

One consequence of the finalized branch is that the startup
time for interactive usage of PyPy is reduced by some 50-70 %. 
This is mainly due to the now-lazy sys and builtin module. 
Btw, if you access "__builtin__.__dict__" or "sys.__dict__" 
then the lazyness ends: all attributes will be loaded and
you get a normal dictionary.  So from application level 
the module-lazyness is basically invisible. 

There are still bits and pieces to be further cleaned 
or fixed up but that can happen in the trunk now. 



git-svn-id: http://codespeak.net/svn/pypy/dist@9332 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada