Trying to combine the typer and genc, we get a bit more mess into genc_funcdef but it is hopefully still reasonable.
Types are a bit difficult to follow right now; they don't show up in the
pygame viewers, and the chosen local variable names don't hint at them either.  
Debugging segfaults might become a problem because type errors (which are real
errors) are lost in the sea of the C compiler's warnings (which we don't care
about, like label not used).


git-svn-id: http://codespeak.net/svn/pypy/dist@10236 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada