added graph pages for only the class hierarchy and a localized call graph for functions (showing just a single function and its direct callers and callees, there are extra nodes that give access to the same for those)
there's a threshold for translate_pypy.py for the number of functions in translator set with -huge=.
Above the threshold we don't try to show the full call graph but only the localized one for the entry-point.

Command are accesible at the translate_pypy prompt to ask for a graph: see help graphs:

- for example classhier will show the class hierarchy graph,

- try showg intobject.W_IntObject.__mm_mul after translate_pypy targetpypy1 analysis for an
example of localized call graph.



git-svn-id: http://codespeak.net/svn/pypy/dist@10294 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada