translated almost all possible code pieces to interp level.
reversed the change to applevelinterp in the source files.

applevel is now a function that figures out whether it can
translate the code or not.
All code snippets have been inspected for translatability
and marked with NOT_RPYTHON if necessary.
Some can be easily enabled, some are hard (using yield)

git-svn-id: http://codespeak.net/svn/pypy/dist@10351 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada