added caching of translated snippets. The cache file is currently pypy.interpreter._interplevel_cache.py . Name and position are arguable,but easy to change -- feel free!
After a single run of test_all.py, the cache file is filled, and
flow space is no longer needed. The speed boost for test_all is
quite noticeable, appears to be 30-40 %.

A side effect of caching is that tests *might* behave different,
depending on the cache file being filled or not. To be safe,
the cache can be cleared before testing. A simple double-click
on the cache file makes it self-destruct.

The cache does very simple heuristics: Every unseen source text
gets its new entry. There is a simple version check against
GI_VERSION in geninterplevel.py, which I think is sufficient.


git-svn-id: http://codespeak.net/svn/pypy/dist@10503 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada