Reworked the implicit exception handling.  Now, an implicit exception is raised as an ImplicitOperationError (a subclass of OperationError). If it is caught, fine.  If it falls all the way through the function, then it means it is not caught, and we now it should be ignored.  If it is caught but re-raised with a bare 'raise' statement, a minor hack ensures that a normal OperationError gets re-raised then.
For convenience, when an ImplicitOperationError falls through, the
flow obj space makes a link to the except block but with a constant
AssertionError, basically saying "this implicit exception shouldn't
have happened".  It might be nice to try running code with such
assertion checks.

Normally, though, we don't want them.  So the flow graph model is now
that such jumps with AssertionError to the except block are regarded as
"dead code", i.e. links that are not supposed to be ever followed at
run-time.  So simplify.py now removes such links.

Some carefulness is needed to avoid breaking the graph by removing too
many links.

New tests check that the correct implicit exceptions are removed and
not more.


git-svn-id: http://codespeak.net/svn/pypy/dist@11439 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada