issue26 resolved
Got rid of getthreadlocals() and the strange interactions with execution
contexts that disappear if there is no frame left in the frame stack.
Now each space creates a single ExecutionContext and saves it in a private
attribute.  (The attribute is private to continue to force the use of
space.getexecutioncontext(), which in the future should be modified to
cache and return one ExecutionContext per thread.)

Fixed a large number of small API functions that don't need so many
arguments any more, e.g. enter()/leave().

Bumped geninterplevel's version number -- cache issues detected...



git-svn-id: http://codespeak.net/svn/pypy/dist@12365 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada