issue51: in-progress
sneaking in a change that helps explaining things in the
documentation (and also makes PyPy more introspectable):
mixed modules now have a __file__ attribute that points
to the underlying __init__.py file that contains the
name -> file/object mappings.  

I actually also intend to rename LazyModule -> MixedModule because the
lazyness is just an implementation detail (currently, it may be possible
to get it RPythonic enough to translate it but currently it just gets
completely unlazified for translation).   



git-svn-id: http://codespeak.net/svn/pypy/dist@12432 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada