issue18 testing
Up-to-dated architecture.txt, and added links for further reading
to the relevant documents at the end of each section.  We no longer
talk in any detail about the standard object space or the trace
object space in architecture.txt now; this has been moved to
objspace.txt (with links).

Added a table of contents and a better section/subsection organization.



git-svn-id: http://codespeak.net/svn/pypy/dist@12566 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada