issue68 resolved issue37 testing
cleaned up the LLVM refs (and all the other optional 
tools refs, added the py lib), added a more prominent 
note and am considering the LLVM issue closed.  

The only issue i have with issue37 (getting-started) is that
iam not sure if it wouldn't be better to move the "directory
reference" to the main doc page and reference it from "where
to start reading sources".   I tend to do this now. 



git-svn-id: http://codespeak.net/svn/pypy/dist@12609 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada