issue39 resolved
Brain-stormed short mission statement for the front page.  We tried to make it
very self-contained, and very short.  Now a link, "more...", points to an 
extended version which gives the main goals.  The links to the existing 
projects -- mainly Psyco and Stackless -- are moved in the next paragraph,
as well as to the Related projects section.



git-svn-id: http://codespeak.net/svn/pypy/dist@12642 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada