Revised specialization, see annotation/specialize.py and changes in bookkeeper;
now annotator can receive a policy object to specify both overrides and specializations through the methods

specialize and override

a base class version is defined in annotation/policy.py, it will use methods of the form specialize__xxx or override__xxx
chosen by attaching to the callable in user-code the attribute _annspecialcase_ with values:

"specialize:xxx"
"override:xxx"

respectively.

ann_override now defines such a policy class which is used by translate_pypy,

(Notice: if a policy is not passed to the annotator an even more basic policy without any overrides nor specializations is used.)

- 'override' methods take the policy and inputcells

- 'specialize'  methods take the policy and a bookkeeper, the involved spaceop, the involved callable func, 
an Arguments object with SomeObjects and a mono flag whether just one statically known callable is involved;

they should normally return a _tuple_ key to identify the specialized version (which if absent will be created), or None meaning no 
specialization

the policy can define a default_specialize that is always invoked, the version in the base class does the specializion for *arg 
functions

There's no general 'location' specialization anymore, for classes there's "specialize:ctr_location" which is implemented
in specialize.py and exposed by the base policy class.

Now phases successive to annotation can query whether a call space op involves specialized callables:

bookkeeper.query_spaceop_callable(spaceop) can be called with a call spaceop that has the called obj annotated as some PBC
it will return possibly a new PBC with the specialized version effectively considered by the annotator for the callable
(it also returns a flag which is normally false unless, in the only case (for now), the callable involved was marked 
"specialize:memo")

the call site families and tables now refer to specialized versions of callables, not the original ones.





git-svn-id: http://codespeak.net/svn/pypy/dist@13088 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada