support for runtime type info (as unique pointers to special opaque objects) for GcStructs, it is possible to attach a query function to GcStruct type that will be used to discover the runtime type info at runtime.
attachRuntimeTypeInfo(GSTRUCT) activates the support for GCSTRUCT, optionally the function as functionptr with signature
(Ptr(GCSTRUCT)) -> Ptr(RuntimeTypeInfo) can be passed.

getRuntimeTypeInfo(GCSTRUCT) returns the statically associated runtime type info object of GCSTRUCT,

runtime_type_info(p) is runtime invocable and returns the runtime type info for p, assuming p it's a pointer to a GcStruct.

This support will be used to implement proper deallocation for pointers to inlined GcStructs.

  


git-svn-id: http://codespeak.net/svn/pypy/dist@13449 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada