Hack to support the '@' token on top of Python 2.3.
Some notes about the pyparser:  I realize that it was a quick hack, but we
should keep in mind that at some point we need something more clearly
independent of CPython's interp-level standard library 'symbol' and 'token'
modules.  I suppose that it should also be refactored to isolate in non-global
instances the tokens and symbols to use, in the same way as the grammar is
isolated.



git-svn-id: http://codespeak.net/svn/pypy/dist@14389 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada