Almost all exceptions tests now pass.
One, possible PBC related, test is failing.

Two tests are failing because an exception in the llvm code is currently
not converted to a CPython exception. The imported (pyrex) module for some
reason then exits but I am not sure if I want the put any effort into
fixing this (at the moment). Especially since the purpose is to have a
selfreliant 'llvm-executable'.



git-svn-id: http://codespeak.net/svn/pypy/dist@15510 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada