issue106 in-progress
start of refactoring of the implementation of identity hash computation. Use the rtyper support for hash() and
value preservation across translation.

This is a first stab. By attaching a identity_hash using hash() only where stricly necessary we can avoid
all W_Root object to internally grow an hash_cache field: this rearragement is to be done later.



git-svn-id: http://codespeak.net/svn/pypy/dist@15583 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada