Wild goose chase: introduced a test to discover why graphs seem to not be in SSA form (or what SSA form means here http://en.wikipedia.org/wiki/SSA_form) when running translate_pypy.
Ironically the problem stemmed from SSI_to_SSA() transformation being run
inadvertently via backend_optimizations from translate_pypy.

ASIDE - any pointers to any docs relating to acronyms used in the
transformation would be much appreciated.  Before I read the docstring I
assumed it was doing the opposite of what it does.  Sorry for my ignorance in
this matter.

The test was a interesting so checking it in and also reenabled a test which
was related to this.



git-svn-id: http://codespeak.net/svn/pypy/dist@16054 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada