The 'thread' module implemented with OS threads and a Global Interpreter Lock (as in CPython).
Must be selected with --usemodules=thread.

Very occasionally, one test fails.  Hard to figure out what's wrong :-(

The GIL itself is further separated from the rest of the code, which is about
using OS threads and locks.  It's almost completely isolated in gil.py.

The 'thread._local' class is much nicer than in CPython: thanks to the
getdict() method on all w_ objects, we can do the "right thing" and have
getdict() return a dictionary that depends on the current thread.  No need to
patch the __dict__ of the thread._local instance in __getattribute__,
__setattr__ and __delattr__.

Experimentally, the 'thread' module comes with its own rpython support in the
'rpython' submodule: it provides rpython support for interp-level
thread.start_new_thread(), thread.get_ident(), thread.allocate_lock() and lock
objects (the latter using the "external types" of the previous check-in).



git-svn-id: http://codespeak.net/svn/pypy/dist@16174 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada