(rxe, arigo)
Support in genc for prebuilt external objects -- in this case, prebuilt locks.
Split OpaqueNode into xxx_OpaqueNode classes for the various usages of opaque
prebuilt constants.

Lots of changes; this required producing and calling start-up code to the
generated C source.

Each thread_*.h needs to provide in the RPyOpaque_INITEXPR_ThreadLock macro an
expression suitable for a lock initializer (though RPyThreadLockInit() is
still called on the structure from the start-up code).



git-svn-id: http://codespeak.net/svn/pypy/dist@16302 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada