(hpk, pedronis)
fixing a couple of problems with the compiler package (XXX port to _stablecompiler) 
this is almost too involved to be described in detail right now. 
Basically we are trying to be more careful and specific 
regarding determining if we are in an optimized scope or not
and also about forcing names in nested scopes to become
globals. 

added new tests that now pass (and some in test_scope should be fixed by that)



git-svn-id: http://codespeak.net/svn/pypy/dist@16416 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada