added os.remove and os.unlink. This was missing in order to run tests with the compiled pypy.
Also started to add the proper docstrings.
I'm not sure whether we should do this or fetch them
from the according CPython functions?

A remaining blocker that still hinders me from testing
is that the compiler seems to crash the compiled PyPy.
Try for instance the file test_complex.
If you put the whole big class into a triple quoted string ('''),
it at least tries to do something. Otherwise, the session
immediately crashes, whether interactive or not!


git-svn-id: http://codespeak.net/svn/pypy/dist@16439 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada