worked hard on making the compiled pypy do testing.
This had the side effect of many fixes, so it was not lost.

But there are still problems.

First thing: when I use the half-faked compiler to compile test_complex,
pypy crashes with a core dump.
For that reason, I also added a mode to fake the complete compiler.
That finally works, but then _sre is missing.

We need to decide if we want to continue with this right now.

git-svn-id: http://codespeak.net/svn/pypy/dist@16580 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada