Some tentative clean-ups of conftest.py w.r.t. object space creation and caching.  Now the gettestobjspace() function takes an optional name 'std'/'thunk' as before, but also keyword options to pass to the space. The result is cached.
As far as I can tell there was already no difference left between
getobjspace() and gettestobjspace().

For interp-level test classes, 'cls.space' attribute is now lazily computed
the first time the attribute is actually accessed.  This is cool when working
on parts of PyPy that don't need an object space around, or that only need a
custom one (e.g. with a usemodules keyword).



git-svn-id: http://codespeak.net/svn/pypy/dist@17098 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada