this check-in is just to show a funny observation.
When a variable is conditionally assigned two different ranges,
the ranges are kept, unless the steps are different.
Well, after all this is not surprizing.
But I thing the action taken is not pleasant:
the range is turned into a list in this case!

I guess what we want is either special-casing the respective blocks,
or at least map this to the variable step case, if this is possible.

Please let me know what you think.


git-svn-id: http://codespeak.net/svn/pypy/dist@17487 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada