Integrated a simple malloc-removing algorithm.  This version is based on http://codespeak.net/svn/user/arigo/hack/pypy-hack/remove_malloc2.py.
Given a graph, it divides the vars in families, where two vars are in the same
family if a value can go from one to the other across links (or same_as or
cast_pointer operations); the family records all usages of its variables.  Then
we just use this information to see which families correspond to
malloc/getfield/setfield only.

The family ("LifeTimes") are probably useful for other kind of analysis as
well, e.g. (samuele) to property detect what exceptions a graph can raise.



git-svn-id: http://codespeak.net/svn/pypy/dist@17505 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada