Please have a good look at bench-unix.py ! I think we have been optimizing with the wrong numbers in our mind.
There were two bugs in the formula's used:
1. We did not take into account that we run less iterations
   in Richards, causing our numbers to be 10x too optimistic
2. The formula for number of pystones was reversed

The good news is that it looks like PyPy is beating CPython handsdown
with regards to number of Pystones.

The other news is that Richards probably has a bottleneck.

(note: bench-windows.py still uses the incorrect formulas!)


but don't worry:
There always is a change that I am completely wrong!



git-svn-id: http://codespeak.net/svn/pypy/dist@17801 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada