Don't use the GC_all_interior_pointers flag (crashes on Windows for initialization-order-related issues).  Use the equivalent way of always mallocing with IGNORE_OFF_PAGE.
For reference, the GC_MALLOC macro variants:

ATOMIC means "don't search in the object for further pointer".  non-ATOMIC
means "search in the complete object for further pointer".  Normally, if there
is anywhere in memory a pointer-looking number that appears to point to any of
the allocated bytes of an object, the object is kept alive.  IGNORE_OFF_PAGE
means "ignore pointer-looking numbers that point inside the object but too far
away from the start of the object".



git-svn-id: http://codespeak.net/svn/pypy/dist@18214 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada