use the complete_callgraph property now.
This has the advantage that the picture gets much more complete.

Sometimes it now gets over-complete, because transitions which
already existed before some transformations, may be found twice, now.

A cleanup may be needed at some point, but for now it is good enough.

See for instance targetrpystonedalone.py, where I have put a recursion
just for testing it with graphics (and actually stackless).
The recursion in pysone_main is shown twice, now.

git-svn-id: http://codespeak.net/svn/pypy/dist@19299 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada