[117 passing, 46 known to fail]
The 46 that are known to fail now have a comment stateing their
probable issue. Many of these have to do that we no longer keep
the types around. 

Furthermore I would like to have the exception raising operations
implemented in Python too. I can't remember why they are required
to be implemented in the backend. But at least genc, genllvm and genjs,
could share them on a somewhat higher level.



git-svn-id: http://codespeak.net/svn/pypy/dist@19724 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada