Start of the hint()-driven approach.  So far, it works on the small example (and only there :-).  The idea is to have a flag 'fixed' on LLRuntimeValues that can be set to True a posteriori, to mean that the constant contained in this LLRuntimeValue should be passed over to the next block as a constant -- instead of the default behavior which is to turn it into a variable.  The 'hint' operation sets some 'fixed' flags to True and raises a RestartCompleting to restart the whole process.  An 'origin' attribute on LLRuntimeValues tracks the multiple possible histories of a variable back to the point(s) where it was constant.
I removed the code for constant propagation with automatic generalization
because the new a-posteriori-constantification logic is confusing enough to
get right without having to additionally worry about not breaking the
const_propagate policy.  I guess we'll have to re-insert this code when things
start to work again.

Added the hint() in the TL source code and implemented the 'hint' operation in
GenC.  The test_jit_tl is disabled because it explodes.



git-svn-id: http://codespeak.net/svn/pypy/dist@21263 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada